// play/chess-tourney-frontend/src/api/model/readBot.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * (title)
 * API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.
 * OpenAPI spec version: 0.0.0
 */

export interface ReadBot {
  user_id: number;
  bot_id: number;
  /** @minLength 3 */
  username: string;
  model_key?: string;
  /** @minimum 1 */
  playouts_per_move: number;
  /**
   * @minimum 0
   * @maximum 20
   */
  exploration_factor: number;
  /**
   * @minimum 0
   * @maximum 5
   */
  virtual_loss: number;
  /**
   * @minimum -3
   * @maximum 3
   */
  contempt: number;
}